small update

+++ src/main/java/entity/Book.java
@@ -1,12 +1,12 @@
 package entity;
 
 public class Book {
-    private int id;
+    private long id;
     private String name;
     private String author;
-    private int readerId;
+    private long readerId;
 
-    public Book(int id, String name, String author, int readerId) {
+    public Book(long id, String name, String author, long readerId) {
         this.id = id;
         this.name = name;
         this.author = author;
@@ -18,7 +18,7 @@ public class Book {
         this.author = author;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
